Fixed bug where non admin users could access project groups pages

+++ src/subprojectgroupsmenu/RemoveMemberFromProjectGroup.java
@@ -22,34 +22,38 @@ public class RemoveMemberFromProjectGroup extends gui.ProjectGroupsMenu {
 	public void doGet(HttpServletRequest request, HttpServletResponse response) throws IOException {
 		if (loggedIn(request)) {
 			HttpSession session = request.getSession();
-			PrintWriter out = response.getWriter();
-			out.print(getPageIntro());
 			int userPermission = (Integer) session.getAttribute("user_permissions");
-			out.append(generateMainMenu(userPermission, request));
-			out.print(generateSubMenu(userPermission));
-			if (request.getParameter("username") == null) {
-				out.print(users.showUsers(db.getUsers()));
-				out.print("<script>$('#addusertogroup').submit(function (e) { e.preventDefault(); var confirmed = confirm(\"\u00C4r du s\u00E4ker?\");if (confirmed) {$(this).submit();}});</script>");
-			} else {
-				User user = db.getUser(Integer.parseInt(request
-						.getParameter("username")));
-				if (user.getProjectGroup() != 0) {
-					HashMap<Integer, Integer> map = new HashMap<Integer, Integer>();
-					map.put(user.getId(), User.ROLE_NO_ROLE);
-					db.setUserRoles(map);
-					db.removeUserFromProjectGroup(user.getId(),
-							user.getProjectGroup());
-					if (db.getUsers(user.getProjectGroup()).isEmpty()) {
-						out.print(db.getProjectGroup(user.getProjectGroup()).getProjectName() + " \u00E4r tom.");
-					} else {
-						out.print(group.showProjectGroup(db.getUsers(user
-								.getProjectGroup())));
-					}
+			if (userPermission == PERMISSION_ADMIN) {
+				PrintWriter out = response.getWriter();
+				out.print(getPageIntro());
+				out.append(generateMainMenu(userPermission, request));
+				out.print(generateSubMenu(userPermission));
+				if (request.getParameter("username") == null) {
+					out.print(users.showUsers(db.getUsers()));
+					out.print("<script>$('#addusertogroup').submit(function (e) { e.preventDefault(); var confirmed = confirm(\"\u00C4r du s\u00E4ker?\");if (confirmed) {$(this).submit();}});</script>");
 				} else {
-					out.print("<script>$(alert(\"Anv\u00E4ndaren har ingen projektgrupp och kan d\u00E4rf\u00F6r inte tas bort ur en\"))</script>");
+					User user = db.getUser(Integer.parseInt(request
+							.getParameter("username")));
+					if (user.getProjectGroup() != 0) {
+						HashMap<Integer, Integer> map = new HashMap<Integer, Integer>();
+						map.put(user.getId(), User.ROLE_NO_ROLE);
+						db.setUserRoles(map);
+						db.removeUserFromProjectGroup(user.getId(),
+								user.getProjectGroup());
+						if (db.getUsers(user.getProjectGroup()).isEmpty()) {
+							out.print(db.getProjectGroup(user.getProjectGroup()).getProjectName() + " \u00E4r tom.");
+						} else {
+							out.print(group.showProjectGroup(db.getUsers(user
+									.getProjectGroup())));
+						}
+					} else {
+						out.print("<script>$(alert(\"Anv\u00E4ndaren har ingen projektgrupp och kan d\u00E4rf\u00F6r inte tas bort ur en\"))</script>");
+					}
 				}
+				out.print(getPageOutro());
+			} else {
+				response.sendRedirect("");
 			}
-			out.print(getPageOutro());
 		} else {
 			response.sendRedirect("LogIn");
 		}
